// src/shared/interfaces/config.interface.ts
export interface IConfig {
  environment: string;
  port: number;
  logLevel: string;
  weatherApiUrl: string;
  weatherApiKey: string;
  newsApiUrl: string;
  newsApiKey: string;
  currencyApiUrl: string;
  currencyApiKey: string;
  holidayApiUrl: string;
  sportsApiUrl: string;
  sportsApiKey: string;
}